requirements.txt
Jinja2==2.11.2
napalm==3.2.0
ncclient==0.6.9
netaddr==0.8.0
netmiko==3.3.0
nornir==3.0.0
nornir-ansible==2020.7.20
nornir-jinja2==0.1.0
nornir-napalm==0.1.1
nornir-netbox==0.1.1
nornir-netmiko==0.1.1
nornir-scrapli==2020.9.16.post1
nornir-utils==0.1.0
paramiko==2.7.2
PyYAML==5.3.1
requests==2.24.0
rich==8.0.0
scrapli==2020.8.28